+++ drawio/controller/viewercontroller.php
@@ -115,7 +115,7 @@ class ViewerController extends Controller
      * @NoCSRFRequired
      */
     public function index($fileId, $shareToken = NULL, $filePath = NULL) {
-        $this->logger->warning("Open: $fileId $shareToken $filePath", array("app" => $this->appName));
+        //$this->logger->warning("Open: $fileId $shareToken $filePath", array("app" => $this->appName));
         if (empty($shareToken) && !$this->userSession->isLoggedIn()) {
             $redirectUrl = $this->urlGenerator->linkToRoute("core.login.showLoginForm", [
                 "redirect_url" => $this->request->getRequestUri()
@@ -204,7 +204,7 @@ class ViewerController extends Controller
 
         }
 
-        if ($this->userId) {            
+        if ($this->userId) {
             $response = new TemplateResponse($this->appName, "viewer", $params);
         } else {
             $response = new PublicTemplateResponse($this->appName, "viewer", $params);